requirements.txt
streamlit>=1.37
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
scipy>=1.13
plotly>=5.22
python-dotenv>=1.0
sentence-transformers>=3.0
nltk>=3.9
wordcloud>=1.9
Pillow>=10.3
sqlalchemy>=2.0
requests>=2.32
